+++ dart/test/scope_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: deprecated_member_use_from_same_package
+
 import 'package:collection/collection.dart';
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/sentry_tracer.dart';
@@ -417,7 +419,6 @@ void main() {
     test('apply context to event', () async {
       final event = SentryEvent(
         tags: const {'etag': '987'},
-        // ignore: deprecated_member_use_from_same_package
         extra: const {'e-infos': 'abc'},
       );
       final scope = Scope(SentryOptions(dsn: fakeDsn))
@@ -442,9 +443,7 @@ void main() {
       expect(updatedEvent?.tags,
           {'etag': '987', 'build': '579', 'page-locale': 'en-us'});
       expect(
-          // ignore: deprecated_member_use_from_same_package
-          updatedEvent?.extra,
-          {'e-infos': 'abc', 'company-name': 'Dart Inc'});
+          updatedEvent?.extra, {'e-infos': 'abc', 'company-name': 'Dart Inc'});
       expect(updatedEvent?.contexts['theme'], {'value': 'material'});
     });