+++ flutter/example/lib/main.dart
@@ -4,31 +4,31 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:math';
 
+import 'package:dio/dio.dart';
+import 'package:feedback/feedback.dart' as feedback;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:logging/logging.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:provider/provider.dart';
+import 'package:sentry_dio/sentry_dio.dart';
 import 'package:sentry_drift/sentry_drift.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
+import 'package:sentry_hive/sentry_hive.dart';
 import 'package:sentry_isar/sentry_isar.dart';
+import 'package:sentry_logging/sentry_logging.dart';
 import 'package:sentry_sqflite/sentry_sqflite.dart';
 import 'package:sqflite/sqflite.dart';
-
 // import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 // import 'package:sqflite_common_ffi_web/sqflite_ffi_web.dart';
 import 'package:universal_platform/universal_platform.dart';
-import 'package:feedback/feedback.dart' as feedback;
-import 'package:provider/provider.dart';
+
 import 'auto_close_screen.dart';
-import 'drift/database.dart';
 import 'drift/connection/connection.dart';
+import 'drift/database.dart';
 import 'isar/user.dart';
 import 'user_feedback_dialog.dart';
-import 'package:dio/dio.dart';
-import 'package:sentry_dio/sentry_dio.dart';
-import 'package:sentry_logging/sentry_logging.dart';
-import 'package:sentry_hive/sentry_hive.dart';
 
 // ATTENTION: Change the DSN below with your own to see the events in Sentry. Get one at sentry.io
 const String exampleDsn =
@@ -474,6 +474,7 @@ class MainScaffold extends StatelessWidget {
                         final entries = feedback.extra?.entries;
                         if (entries != null) {
                           for (final extra in entries) {
+                            // ignore: deprecated_member_use
                             scope.setExtra(extra.key, extra.value);
                           }
                         }